pyproject.toml
[tool.poetry]
name = "01-bases-de-datos"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jupyter = "^1.1.1"
ipykernel = "^6.29.5"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
numpy = "^2.1.1"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
duckdb = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
